requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.4
django-disqus==0.5
django-toolbelt==0.0.1
gunicorn==19.7.1
mock==2.0.0
pbr==3.1.1
psycopg2==2.7.1
pystache==0.5.4
python-decouple==3.1
pytz==2017.2
six==1.10.0
static3==0.7.0
whitenoise==3.3.0
